+++ src/wallet/rest_api/mod.rs
@@ -1465,6 +1465,7 @@ impl RestApi {
     ///
     /// * If `network` not send, return with default network of the coin.
     /// * You can get `network` and `isDefault` in `networkList` of a coin in the response of `Get /sapi/v1/capital/config/getall (HMAC SHA256)`.
+    /// * To check if travel rule is required, by using  `GET /sapi/v1/localentity/questionnaire-requirements` and if it returns anything other than `NIL` you will need update SAPI to `POST /sapi/v1/localentity/withdraw/apply` else you can continue `POST /sapi/v1/capital/withdraw/apply`. Please note that if you are required to comply to travel rule please refer to the Travel Rule SAPI.
     ///
     /// Weight: 900
     ///
@@ -1696,13 +1697,7 @@ impl RestApi {
     ///
     /// This API will return user-specific Travel Rule questionnaire requirement information in reference to the current API key.
     ///
-    /// * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    ///
-    /// Weight: 18000
-    /// Request limit: 10 requests per second
-    /// > * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    /// limit is 180000/second. Response from the endpoint contains header
-    /// key X-SAPI-USED-IP-WEIGHT-1S, which defines weight used by the current IP.
+    /// Weight: 1
     ///
     /// # Arguments
     ///
@@ -1791,11 +1786,57 @@ impl RestApi {
             .await
     }
 
+    /// Deposit History V2 (for local entities that required travel rule) (supporting network) (`USER_DATA`)
+    ///
+    /// Fetch deposit history for local entities that with required travel rule information.
+    ///
+    /// * Please notice the default `startTime` and `endTime` to make sure that time interval is within
+    /// * If both ``startTime`` and ``endTime`` are sent, time between ``startTime`` and ``endTime`` must
+    ///
+    /// Weight: 1
+    ///
+    /// # Arguments
+    ///
+    /// - `params`: [`DepositHistoryV2Params`]
+    ///   The parameters for this operation.
+    ///
+    /// # Returns
+    ///
+    /// [`RestApiResponse<Vec<models::DepositHistoryV2ResponseInner>>`] on success.
+    ///
+    /// # Errors
+    ///
+    /// This function will return an [`anyhow::Error`] if:
+    /// - the HTTP request fails
+    /// - any parameter is invalid
+    /// - the response cannot be parsed
+    /// - or one of the following occurs:
+    ///   - `RequiredError`
+    ///   - `ConnectorClientError`
+    ///   - `UnauthorizedError`
+    ///   - `ForbiddenError`
+    ///   - `TooManyRequestsError`
+    ///   - `RateLimitBanError`
+    ///   - `ServerError`
+    ///   - `NotFoundError`
+    ///   - `NetworkError`
+    ///   - `BadRequestError`
+    ///
+    ///
+    /// For full API details, see the [Binance API Documentation](https://developers.binance.com/docs/wallet/travel-rule/Deposit-History-V2).
+    ///
+    pub async fn deposit_history_v2(
+        &self,
+        params: DepositHistoryV2Params,
+    ) -> anyhow::Result<RestApiResponse<Vec<models::DepositHistoryV2ResponseInner>>> {
+        self.travel_rule_api_client.deposit_history_v2(params).await
+    }
+
     /// Fetch address verification list (`USER_DATA`)
     ///
-    /// Fetch address verification list
+    /// Fetch address verification list for user to check on status and other details for the addresses stored in Address Book.
     ///
-    /// Weight: 10
+    /// Weight: 1
     ///
     /// # Arguments
     ///
@@ -1941,13 +1982,7 @@ impl RestApi {
     ///
     /// Fetch the VASP list for local entities.
     ///
-    /// * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    ///
-    /// Weight: 18000
-    /// Request limit: 10 requests per second
-    /// > * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    /// limit is 180000/second. Response from the endpoint contains header
-    /// key X-SAPI-USED-IP-WEIGHT-1S, which defines weight used by the current IP.
+    /// Weight: 1
     ///
     /// # Arguments
     ///
@@ -1990,16 +2025,11 @@ impl RestApi {
     ///
     /// Fetch withdraw history for local entities that required travel rule.
     ///
-    /// * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
     /// * `network` may not be in the response for old withdraw.
     /// * Please notice the default `startTime` and `endTime` to make sure that time interval is within
     /// * If both `startTime` and `endTime`are sent, time between `startTime`and `endTime`must be less
     ///
-    /// Weight: 18000
-    /// Request limit: 10 requests per second
-    /// > * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    /// limit is 180000/second. Response from the endpoint contains header
-    /// key X-SAPI-USED-IP-WEIGHT-1S, which defines weight used by the current IP.
+    /// Weight: 1
     ///
     /// # Arguments
     ///
@@ -2044,7 +2074,6 @@ impl RestApi {
     ///
     /// Fetch withdraw history for local entities that required travel rule.
     ///
-    /// * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
     /// * `network` may not be in the response for old withdraw.
     /// * Withdrawal made through /sapi/v1/capital/withdraw/apply may not be in the response.
     /// * Please notice the default `startTime` and `endTime` to make sure that time interval is within
@@ -2055,11 +2084,7 @@ impl RestApi {
     /// * `WithdrawOrderId` only support 1.
     /// * If responsible does not include withdrawalStatus, please input trId or txId retrieve the data.
     ///
-    /// Weight: 18000
-    /// Request limit: 10 requests per second
-    /// > * This endpoint specifically uses per second IP rate limit, user's total second level IP rate
-    /// limit is 180000/second. Response from the endpoint contains header
-    /// key X-SAPI-USED-IP-WEIGHT-1S, which defines weight used by the current IP.
+    /// Weight: 1
     ///
     /// # Arguments
     ///